apps/agent-api/pyproject.toml
[project]
name = "agent-api"
version = "0.1.0"
description = "ADK-based agent API service"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "litellm[ollama]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
